requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
fairlearn>=0.10.0
plotly>=5.20.0
google-generativeai>=0.7.0
imbalanced-learn>=0.12.0
openpyxl>=3.1.0
